--- config/providers.php
@@ -1,20 +1,0 @@
-<?php
-$prefixNameSpace = "Uniondrug\\PolicySdk\\Providers\\";
-return [
-    /*
-     * 华夏
-     */
-    'policy:HX' => $prefixNameSpace . "HuaXiaCompanyProvider",
-    /*
-     * 人保
-     */
-    'policy:RB' => $prefixNameSpace . "RenbaoCompanyProvider",
-    /*
-     * 太保
-     */
-    'policy:TB' => $prefixNameSpace . "TaibaoCompanyProvider",
-    /*
-     * 天安
-     */
-    'policy:TA' => $prefixNameSpace . "TiananCompanyProvider",
-];
