+++ src/Configs/sdk.php
@@ -1,0 +1,19 @@
+<?php
+return [
+    /*
+     * 华夏
+     */
+    'Sdk:HX' => "HuaXiaCompanySdk",
+    /*
+     * 人保
+     */
+    'Sdk:RB' => "RenbaoCompanySdk",
+    /*
+     * 太保
+     */
+    'Sdk:TB' => "TaibaoCompanySdk",
+    /*
+     * 天安
+     */
+    'Sdk:TA' => "TiananCompanySdk",
+];
