--- src/InjectableTrait.php
@@ -1,20 +1,0 @@
-<?php
-/**
- * Created by PhpStorm.
- * User: luzhouyu
- * Date: 18/7/26
- * Time: 上午12:33
- */
-
-namespace Uniondrug\PolicySdk;
-
-/**
- * trait InjectableTrait
- * @property \Uniondrug\PolicySdk\Plugins\Logger $logger
- * @property \Uniondrug\PolicySdk\Plugins\ApiResponse $apiRespons
- * @property \Uniondrug\PolicySdk\Plugins\PolicySDK $policySdk
- */
-trait InjectableTrait
-{
-
-}
